NOBUG: Improve build date calculations.
Without this change, the build date for any
release is always assumed to be today (in YYYMMDD format).

But that's not always true, especially for major and minor releases.

So this introduces a new --date option to force any build date and
also modifies the default calculations for major and minor releases moving
from today to next monday (that is when all releases happen, no matter they
are ready in git since some days before).

Note the new --date option has precedence over any calculation.