NOBUG: Make master weeklies under beta/rc to behave as on-demand
Previously it was going straight to next version dev, when that is only
expected after a major release and branching. Now those weeklies
behave exactly like on-demand ones, just adding the trailing "+".

Note we always use on-demand normally so this bug was not really affecting us,
but better have weeklies being correctly updated to avoid errors.