NOBUG: For back-to-dev type avoid build going backward
Just detected that the type-back-to-dev was moving
the build information backwards from "next monday"
(used by majors and minors) to "today" (used by all the rest
of releases).

This is not very important and, in fact, has happened
with previous back-to-dev without anybody realising it, see:

http://git.moodle.org/gw?p=moodle.git;a=commitdiff;h=730d717

With this change, back-to-dev follows the same "next monday"
rule than releases, so build won't go backwards anymore.