MDLSITE-4220 Ensure version.php changes are the branch point.
We need all the changes performed before version.php changes
happen and we both branch from it and tag it as release.

This implies a bit of dark magic, deleting the version commit, adding
the remaining changes and re-introducing it again but should
work without problems. After branching, master can be changed back.