Add automatic support for major-related release types
When the tool is being used for a major-related release type:
  - on-demand
  - beta
  - rc
  - major
  - back-to-dev
  - on-sync

It's automatically detected if we are in a parallel development
period. If that's the case, the correct (next to be released)
branch is picked automatically (of course, it's still possible
to use the --branch option to force any branch).

Note that, under parallel development, the "back-to-dev" release
type doesn't have anything to do, because there isn't any branch
to move back to dev (master is already dev), so it's detected and
prevented now.

Finally, when the "major" is released from a STABLE development
branch, like will happen whe we release 3.10... some extra instructions
are printed, about the need to create the next parallel development
branch (MOODLE_311_STABLE) if needed.

With this I think that we can consider the toool 100% ready to
support present and future parallel development periods.