Reconcile prerelease notes and warnings with README (on-sync type).
This adds an "IMPORTANT" warning about the need, the last week
rolling on-sync releases... to disable the queues manager, so it
doesn't continue processing issues after on-sync.

Note this is also in the release process, but scheduled to Monday
after on-sync. Better if we disable it on rolling the last week
of on-sync.

Grrr, also detected a typo (bad comparison operator), fixing it here