Fix problem when calculating the name of lang strings install branches
There was a problem when calculating the name of the lang strings
install branches affecting to DEV branches (for STABLE branches it
was working ok).

That's the reason MOODLE_310_STABLE did the work ok and it
correctly found the install_310_STABLE branch.

But MOODLE_311_STABLE, that is a DEV branch (despite its name) was
calculating the name like: install_MOODLE_311_STABLE that doesn't
exist.

We were missing to remove that "MOODLE_" from the name. Never noticed
this for master (our other DEV branch) where it works ok, because
it doesn't have any MOODLE_ as part of the name :-).