Add the new mimetex.linux.aarch64 to be handled as binary /executable
In MDL-70627 we introduced a new mimetex binary, and it was added with
755 permissions (I remember it pretty well because I amended the
original commit to add the executable bits).

But that new file was not known to our beloved prerelease stuff, so
perms were changed back to 644. See 3122b30050c639d4a13a079a52779ec033fcf8a1

This commit just adds the new binary, and also sorts them alphabetically
for easier reading.