Avoid any branch change on beta/rc/major
With the release of 4.0beta it has been detected that
we were re-calculating the branch on beta/rc/major
releases, and it's not needed at all (the branch never
changes on those releases).

So this commit just keeps it unmodified for those releases.

Also, a small change has been performed to the beta regexp
in order to avoid creating 4.0betabeta (double betas). This
doesn't happen ever (we only release ONE beta), but testing
discovered that we were concatenating incorrectly.