Remove the Travis automatisms from major releases
Part of the prerelease.sh script for majors we
were adding a couple of "NOBUG" commits adjusting
the new branch name being created within the
.travis.yml file.

Because in MDLSITE-7135 it has been agreed to drop
support for Travis in core, that automatism is not
needed any more.

I've left a big comment explaining how to add any
new automatism that could arrive in the future
and pointing how we were doing that for the
Travis commits. Maybe we won't need it ever but...

You can verify how it was working if you look, in master,
for "NOBUG:" commits around new .0 major versions.

With this changes, those 2 "NOBUG:" (Travis related) commits
won't be created any more. Can be tested easily by
pre-releasing a --type major version and check that
everything works ok and that there aren't any .travis.yml
modifications in the commits.

Note: If tested... don't forget to remove the just created
MOODLE_402_STABLE branch created in your local clone after
the tests. ;-)