Move from master to main everywhere
Not much to say, the changes are just s/master/main changes
and everything should start working with main without troubles.

Note that the next commit is the one that will be in charge
of ensuring that we still keep pushing to master the same changes
that to main. Look to it for more details.