Ensure that master is updated to main and pushed upstream
With this commit we ensure that any prepared release in the
new main is going to be applied (branch -f) also to master and,
that when releasing... both main and master changes are going to
be sent to moodle.git repos.

Note that it also includes an interim change about making
the main pre-release to pick the install lang changes from
the install_master branch of AMOS. Once the new install_main
branch is created in AMOS, we'll remove that change.

Finally, note that, once we remove the master branches from all
repos... this commit will be reverted 100% (already annotated
in the actions to perform when that happens).