Change var names in bump_dev_ensure_higher()
The terms "major" and "minor" are confusing when talking
about versions, aka: $version = 'YYYYMMDD00.ZZ'; because
they can be confused with "major" and "minor" releases
when reviewing the code.

This just changes, in the bump_dev_ensure_higher() function:
- major => versionint (integer part of the version).
- minor => versiondec (decimal part of the version).

So, it's a simple variable rename and that's all, for clarity.
Tracked @ MDLSITE-7209