Change some variables from "major & minor" to "integer & decimal"
The terms "major" and "minor" are confusing when talking
about versions, aka: $version = 'YYYYMMDD00.ZZ'; because
they can be confused with "major" and "minor" releases
when reviewing the code.

This just changes, along the script, the following variables:
    - versionmajorcurrent =>  integerversioncurrent
    - versionminorcurrent =>  decimalversioncurrent
    - versionmajornew =>  integerversionnew
    - versionminornew =>  decimalversionnew

Also, one regular expression has got its placeholders "major"
and "minor" replaced by "integer" and "decimal".

So, it's a simple variable rename and that's all, for clarity.
Tracked @ MDLSITE-7209