Prevent prerelease.sh near publish time
See MDLSITE-7681 for more details

Basically this prevents `prerelease.sh` to be executed
when we are 60 minutes (configurable) before the packages
publishing happen in the server at 00:50 UTC (configurable).

That's to avoid any problem with both running together
and potentially leading to some weekly versions not being
ever published.

So the script now detects if the server is close to run
the publishing scripts and exits with error, recommending to try
again once the server has finished.