Prevent release.sh if publish happened between commit and now
See MDLSITE-7681 for more details. This completes #92

This is the final guard, on release.sh time, to ensure that we
don't send to the upstream repositories too old commits, when
packaging has already happened after the commits happened, because
that will lead to unpublished weeklies.

Note that we shouldn't arrive to this situation ever, as far as
prerelease.sh is already warning us if we try to create those commits
the hour before packaging happens in the server. So chances of
this second and final check in release.sh to happen are really low.

Only if the integrator pre-releases and then forget to release over
some good hours or things like that.

In any case, if this is detected, the solution is to:
- Rewrite integration.git (all affected branches).
- Proceed to pre-release and release again.

(surely it would be enough to amend the commits to current time
but I wouldn't tell you that, heh)