Some operations can fail if the date/time is not the expected one
Basically, when using english, it works ok (en_US, en_AU.UTF-8, ...)
but can fail with some collations like es_ES.UTf-8 when making
some calculations because the expected formats are different from
what the collation provides.

So, we are going to use LC_ALL=C in all the "date" commands
used by the scripts in order to provide a consistent behaviour
no matter which collation is being used by default.

We could also have set LC_TIME, LANG... but as far as LC_ALL is
the one with max precedence, we just set that one, and only
for the execution of the "date" commands. And "C", because that
is the default, basic collation that all the systems support
much like the POSIX one, only ASCII and enough for our interests).